+++ packages/http/src/data-http-client.module.ts
@@ -6,6 +6,7 @@ import { DATA_REQUEST_OPTIONS_CONFIG } from './configs/data-request-options.conf
 import { DataConfiguratorService } from './services/data-configurator.service';
 import { DataHttpClient } from './services/data-http.client';
 import { DefaultHttpClientInterceptor } from './services/default-http-client-interceptor';
+import { LimitConcurrencyService } from './services/limit-concurency.service';
 import { DATA_CONFIG_SERVICE_TOKEN } from './tokens/data-config-service.token';
 import { DATA_HTTP_CLIENT_INTERCEPTOR } from './tokens/data-http-client-interceptor.token';
 
@@ -22,6 +23,7 @@ export class DataHttpClientModule {
                 { provide: DATA_CONFIG_SERVICE_TOKEN, useValue: { ...DATA_REQUEST_OPTIONS_CONFIG, ...options } },
                 { provide: DATA_HTTP_CLIENT_INTERCEPTOR, useClass: DefaultHttpClientInterceptor },
                 DataHttpClient,
+                LimitConcurrencyService,
                 ...clients
             ]
         };